♻️ refactor(utils): 将 MQTT 会话边界移回 utils 层
